services/memory/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
asyncpg==0.30.0
sentence-transformers==3.3.1
faiss-cpu==1.9.0
numpy==1.26.4
pydantic==2.10.3
pydantic-settings==2.6.1
prometheus-client==0.21.0
